+++ recipes/lets_encrypt_cert.rb
@@ -12,7 +12,7 @@ end
 package 'python-certbot-apache'
 
 command = "certbot --apache certonly -n --email #{cert_email} --agree-tos"
-names = generate_alt_names
+names = generate_domain_names
 names.each do |name|
   command += " -d #{name}"
 end
@@ -22,4 +22,10 @@ puts("CERTBOT COMMAND: #{command}")
 bash 'Get Lets Encrypt Certificate' do
   code command
   action :nothing
+end
+
+file 'Certbot Record' do
+  path '/opt/chef/run_record/certbot_command.txt'
+  content command
+  notifies :run, 'bash[Get Lets Encrypt Certificate]', :immediate
 end